fix: Stdio example server works with MCP inspector
This change requires the fixes made in cced0e4.

The previous stdio example server fails to even load under MCP Inspector
and gives an inscrutable connection error that requires loading the
developer console to understand. This is due to a number of issues,
including not using camelCase for parameters in the initialize request.
Not supporting the method "tools/list". Not supporting "tools/call". And
last, not returning the result of a tool call in the expected format.

With these changes, the example server now works with MCP inspector.

1. The initialize response returns fields in camelCase and returns the
   correct object for tools.
2. Handle "tools/list" to return the available tools.
3. Fixup the match to correctly handle "tools/call"
4. Fixup the object returned by tool handlers.

I did not fix the client nor did I modify the SSE examples. I only
tested this using MCP inspector. To making testing this easy, I wrote a
script I called `test-it.sh` that lives alongside my mcpr checkout. My
test-it script looks like this:

    #!/usr/bin/env bash

    pushd mcpr
    cargo install --path .
    popd

    # Uncomment this line if you want a "clean" build
    # Since the script will overwrite files, you can get faster iteration
    times
    # by leaving this directory in place
    # rm -rf my-stdio-project
    mcpr generate-project --name my-stdio-project --transport stdio

    cd my-stdio-project
    ./mcp-inspector.sh

My edit-compile-debug loop basically was me editing the server template
string and then running the following workflow until it worked:

1. Start MCP inspector using my `test-it.sh` script.
2. Loading http://localhost:5173/ in my browser.
3. Clicking the connect button.
4. Clicking the tools button.
5. Clicking list tools.
6. Clicking the "hello" tool.
7. Entering my name in the "name" box and hitting "Run tool"
8. Observing that MCP Inspector prints "Tool Result: Success" with the
   expected output from the tool call.

If any of the above steps failed, debug, fix and retry.